+++ R/gg_bin_2d.R
@@ -435,7 +435,7 @@ gg_bin_2d <- function(
   }
 
   if (!col_scale) {
-    if (rlang::is_null(pal)) pal <- "#36648b"
+    if (rlang::is_null(pal)) pal <- "#357BA2"
     else pal <- as.vector(pal[1])
 
     plot <- plot +
@@ -1330,7 +1330,7 @@ gg_bin_2d <- function(
         else col_breaks <- scales::breaks_pretty(n = 5)
       }
 
-      if (rlang::is_null(pal)) pal <- viridisLite::mako(n = 20, direction = -1)
+      if (rlang::is_null(pal)) pal <- viridisLite::mako(n = 18, direction = -1)
 
       if (rlang::is_null(col_labels)) {
         if (col_numeric | col_null) {