+++ R/gg_ribbon.R
@@ -84,10 +84,10 @@
 #'     x = year,
 #'     ymin = level_min,
 #'     ymax = level_max,
-#'     pal = scales::alpha("steelblue4", 0),
+#'     pal = scales::alpha("steelblue", 0),
 #'     y_title = "Level"
 #'   ) +
-#'   geom_line(aes(x = year, y = level), col = "steelblue4")
+#'   geom_line(aes(x = year, y = level), col = "steelblue")
 #'
 gg_ribbon <- function(
     data = NULL,
@@ -496,7 +496,7 @@ gg_ribbon <- function(
   }
 
   if (!col_scale) {
-    if (rlang::is_null(pal)) pal <- "#36648b"
+    if (rlang::is_null(pal)) pal <- "#357BA2"
     else pal <- as.vector(pal[1])
 
     plot <- plot +
@@ -1391,7 +1391,7 @@ gg_ribbon <- function(
         else col_breaks <- scales::breaks_pretty(n = 5)
       }
 
-      if (rlang::is_null(pal)) pal <- viridisLite::mako(n = 20, direction = -1)
+      if (rlang::is_null(pal)) pal <- viridisLite::mako(n = 18, direction = -1)
 
       if (rlang::is_null(col_labels)) {
         if (col_numeric | col_null) {